pyproject.toml
[project]
name = "splat_ready"
version = "1.0.0"
description = "SplatReady - convert video to COLMAP datasets ready for Gaussian splat training"
author = "Jacob van Beets"

dependencies = [
    "Pillow",
    "piexif",
]

[tool.lichtfeld]
auto_start = true
hot_reload = true
plugin_api = ">=1,<2"
lichtfeld_version = ">=0.4.2"
required_features = []
